rm union type
