+++ API.md
@@ -119,8 +119,18 @@ const cloudWatchLogsOutputOptions: CloudWatchLogsOutputOptions = { ... }
 
 | **Name** | **Type** | **Description** |
 | --- | --- | --- |
+| <code><a href="#image-scanner-with-trivy.CloudWatchLogsOutputOptions.property.type">type</a></code> | <code><a href="#image-scanner-with-trivy.ScanLogsOutputType">ScanLogsOutputType</a></code> | *No description.* |
 | <code><a href="#image-scanner-with-trivy.CloudWatchLogsOutputOptions.property.logGroupName">logGroupName</a></code> | <code>string</code> | *No description.* |
-| <code><a href="#image-scanner-with-trivy.CloudWatchLogsOutputOptions.property.type">type</a></code> | <code>string</code> | *No description.* |
+
+---
+
+##### `type`<sup>Required</sup> <a name="type" id="image-scanner-with-trivy.CloudWatchLogsOutputOptions.property.type"></a>
+
+```typescript
+public readonly type: ScanLogsOutputType;
+```
+
+- *Type:* <a href="#image-scanner-with-trivy.ScanLogsOutputType">ScanLogsOutputType</a>
 
 ---
 
@@ -128,16 +138,6 @@ const cloudWatchLogsOutputOptions: CloudWatchLogsOutputOptions = { ... }
 
 ```typescript
 public readonly logGroupName: string;
-```
-
-- *Type:* string
-
----
-
-##### `type`<sup>Required</sup> <a name="type" id="image-scanner-with-trivy.CloudWatchLogsOutputOptions.property.type"></a>
-
-```typescript
-public readonly type: string;
 ```
 
 - *Type:* string
@@ -453,6 +453,176 @@ Put each line you write in the file into one element of the array.
     aws-account-id
 ```
 
+
+### ScanLogsOutputOptions <a name="ScanLogsOutputOptions" id="image-scanner-with-trivy.ScanLogsOutputOptions"></a>
+
+Output configurations for scan logs.
+
+#### Initializer <a name="Initializer" id="image-scanner-with-trivy.ScanLogsOutputOptions.Initializer"></a>
+
+```typescript
+import { ScanLogsOutputOptions } from 'image-scanner-with-trivy'
+
+const scanLogsOutputOptions: ScanLogsOutputOptions = { ... }
+```
+
+#### Properties <a name="Properties" id="Properties"></a>
+
+| **Name** | **Type** | **Description** |
+| --- | --- | --- |
+| <code><a href="#image-scanner-with-trivy.ScanLogsOutputOptions.property.type">type</a></code> | <code><a href="#image-scanner-with-trivy.ScanLogsOutputType">ScanLogsOutputType</a></code> | *No description.* |
+
+---
+
+##### `type`<sup>Required</sup> <a name="type" id="image-scanner-with-trivy.ScanLogsOutputOptions.property.type"></a>
+
+```typescript
+public readonly type: ScanLogsOutputType;
+```
+
+- *Type:* <a href="#image-scanner-with-trivy.ScanLogsOutputType">ScanLogsOutputType</a>
+
+---
+
+### ScannerCustomResourceProps <a name="ScannerCustomResourceProps" id="image-scanner-with-trivy.ScannerCustomResourceProps"></a>
+
+Lambda function event object for Scanner Custom Resource.
+
+#### Initializer <a name="Initializer" id="image-scanner-with-trivy.ScannerCustomResourceProps.Initializer"></a>
+
+```typescript
+import { ScannerCustomResourceProps } from 'image-scanner-with-trivy'
+
+const scannerCustomResourceProps: ScannerCustomResourceProps = { ... }
+```
+
+#### Properties <a name="Properties" id="Properties"></a>
+
+| **Name** | **Type** | **Description** |
+| --- | --- | --- |
+| <code><a href="#image-scanner-with-trivy.ScannerCustomResourceProps.property.addr">addr</a></code> | <code>string</code> | *No description.* |
+| <code><a href="#image-scanner-with-trivy.ScannerCustomResourceProps.property.exitCode">exitCode</a></code> | <code>number</code> | *No description.* |
+| <code><a href="#image-scanner-with-trivy.ScannerCustomResourceProps.property.exitOnEol">exitOnEol</a></code> | <code>number</code> | *No description.* |
+| <code><a href="#image-scanner-with-trivy.ScannerCustomResourceProps.property.ignoreUnfixed">ignoreUnfixed</a></code> | <code>string</code> | *No description.* |
+| <code><a href="#image-scanner-with-trivy.ScannerCustomResourceProps.property.imageConfigScanners">imageConfigScanners</a></code> | <code>string[]</code> | *No description.* |
+| <code><a href="#image-scanner-with-trivy.ScannerCustomResourceProps.property.imageUri">imageUri</a></code> | <code>string</code> | *No description.* |
+| <code><a href="#image-scanner-with-trivy.ScannerCustomResourceProps.property.platform">platform</a></code> | <code>string</code> | *No description.* |
+| <code><a href="#image-scanner-with-trivy.ScannerCustomResourceProps.property.scanners">scanners</a></code> | <code>string[]</code> | *No description.* |
+| <code><a href="#image-scanner-with-trivy.ScannerCustomResourceProps.property.severity">severity</a></code> | <code>string[]</code> | *No description.* |
+| <code><a href="#image-scanner-with-trivy.ScannerCustomResourceProps.property.trivyIgnore">trivyIgnore</a></code> | <code>string[]</code> | *No description.* |
+| <code><a href="#image-scanner-with-trivy.ScannerCustomResourceProps.property.output">output</a></code> | <code><a href="#image-scanner-with-trivy.ScanLogsOutputOptions">ScanLogsOutputOptions</a></code> | *No description.* |
+
+---
+
+##### `addr`<sup>Required</sup> <a name="addr" id="image-scanner-with-trivy.ScannerCustomResourceProps.property.addr"></a>
+
+```typescript
+public readonly addr: string;
+```
+
+- *Type:* string
+
+---
+
+##### `exitCode`<sup>Required</sup> <a name="exitCode" id="image-scanner-with-trivy.ScannerCustomResourceProps.property.exitCode"></a>
+
+```typescript
+public readonly exitCode: number;
+```
+
+- *Type:* number
+
+---
+
+##### `exitOnEol`<sup>Required</sup> <a name="exitOnEol" id="image-scanner-with-trivy.ScannerCustomResourceProps.property.exitOnEol"></a>
+
+```typescript
+public readonly exitOnEol: number;
+```
+
+- *Type:* number
+
+---
+
+##### `ignoreUnfixed`<sup>Required</sup> <a name="ignoreUnfixed" id="image-scanner-with-trivy.ScannerCustomResourceProps.property.ignoreUnfixed"></a>
+
+```typescript
+public readonly ignoreUnfixed: string;
+```
+
+- *Type:* string
+
+---
+
+##### `imageConfigScanners`<sup>Required</sup> <a name="imageConfigScanners" id="image-scanner-with-trivy.ScannerCustomResourceProps.property.imageConfigScanners"></a>
+
+```typescript
+public readonly imageConfigScanners: string[];
+```
+
+- *Type:* string[]
+
+---
+
+##### `imageUri`<sup>Required</sup> <a name="imageUri" id="image-scanner-with-trivy.ScannerCustomResourceProps.property.imageUri"></a>
+
+```typescript
+public readonly imageUri: string;
+```
+
+- *Type:* string
+
+---
+
+##### `platform`<sup>Required</sup> <a name="platform" id="image-scanner-with-trivy.ScannerCustomResourceProps.property.platform"></a>
+
+```typescript
+public readonly platform: string;
+```
+
+- *Type:* string
+
+---
+
+##### `scanners`<sup>Required</sup> <a name="scanners" id="image-scanner-with-trivy.ScannerCustomResourceProps.property.scanners"></a>
+
+```typescript
+public readonly scanners: string[];
+```
+
+- *Type:* string[]
+
+---
+
+##### `severity`<sup>Required</sup> <a name="severity" id="image-scanner-with-trivy.ScannerCustomResourceProps.property.severity"></a>
+
+```typescript
+public readonly severity: string[];
+```
+
+- *Type:* string[]
+
+---
+
+##### `trivyIgnore`<sup>Required</sup> <a name="trivyIgnore" id="image-scanner-with-trivy.ScannerCustomResourceProps.property.trivyIgnore"></a>
+
+```typescript
+public readonly trivyIgnore: string[];
+```
+
+- *Type:* string[]
+
+---
+
+##### `output`<sup>Optional</sup> <a name="output" id="image-scanner-with-trivy.ScannerCustomResourceProps.property.output"></a>
+
+```typescript
+public readonly output: ScanLogsOutputOptions;
+```
+
+- *Type:* <a href="#image-scanner-with-trivy.ScanLogsOutputOptions">ScanLogsOutputOptions</a>
+
+---
 
 ## Classes <a name="Classes" id="Classes"></a>
