+++ src/image-scanner-with-trivy.ts
@@ -13,12 +13,6 @@ import {
 import { ILogGroup } from 'aws-cdk-lib/aws-logs';
 import { Provider } from 'aws-cdk-lib/custom-resources';
 import { Construct } from 'constructs';
-import {
-  ScanLogsOutputOptions,
-  CloudWatchLogsOutputOptions,
-  ScannerCustomResourceProps,
-  ScanLogsOutputType,
-} from './types';
 
 /**
  * Enum for Severity Selection
@@ -53,6 +47,27 @@ export enum Scanners {
 export enum ImageConfigScanners {
   CONFIG = 'config',
   SECRET = 'secret',
+}
+
+/**
+ * Enum for ScanLogsOutputType
+ */
+export enum ScanLogsOutputType {
+  CLOUDWATCH_LOGS = 'cloudWatchLogs',
+}
+
+/**
+ * Output configurations for scan logs.
+ */
+export interface ScanLogsOutputOptions {
+  readonly type: ScanLogsOutputType;
+}
+
+/**
+ * Output configuration for scan logs to CloudWatch Logs.
+ */
+export interface CloudWatchLogsOutputOptions extends ScanLogsOutputOptions {
+  readonly logGroupName: string;
 }
 
 /**
@@ -109,6 +124,23 @@ class CloudWatchLogsOutput extends ScanLogsOutput {
       logGroupName: this.logGroup.logGroupName,
     };
   }
+}
+
+/**
+ * Lambda function event object for Scanner Custom Resource.
+ */
+export interface ScannerCustomResourceProps {
+  readonly addr: string;
+  readonly imageUri: string;
+  readonly ignoreUnfixed: string;
+  readonly severity: string[];
+  readonly scanners: string[];
+  readonly imageConfigScanners: string[];
+  readonly exitCode: number;
+  readonly exitOnEol: number;
+  readonly trivyIgnore: string[];
+  readonly platform: string;
+  readonly output?: ScanLogsOutputOptions;
 }
 
 export interface ImageScannerWithTrivyProps {
