+++ src/index.ts
@@ -1,2 +1,1 @@
 export * from './image-scanner-with-trivy';
-export { ScanLogsOutputType, CloudWatchLogsOutputOptions } from './types';
